fix: remove sheet nesting
I tried with our own graph on tableau public, as well as the one from the tutorial ( https://public.tableau.com/views/WorldIndicators/GDPpercapita ). In both cases the issue is, that `getActiveSheet()` is already returning the sheet, so there is no method `getWorksheets()`. This is only failing the function, but whenever we try to change the filter, it is crashing the entire component.